Move grid 2d utility functions to spatial_utils.h, update SymWorld to accomodate change
